+++ js/createObject.js
@@ -1,0 +1,23 @@
+import {NAMES,COMMENTS,DESCRIPTIONS,COMMENTS_COUNT,getRandomArrayElement} from './data.js';
+import {getRandomPositiveInteger, getCommentId, createIdGenerator} from './utils.js';
+const createComments = () => Array.from({length: COMMENTS_COUNT},() => ({
+  id: getCommentId(),
+  avatar: `img/avatar-${getRandomPositiveInteger(0, 6)}.svg`,
+  message: getRandomArrayElement(COMMENTS),
+  name: getRandomArrayElement(NAMES),
+}));
+const createPhotos = () =>{
+  const getRandomIdNumber = createIdGenerator();
+  const getRandomUrlNumber = createIdGenerator();
+  const photos = [];
+  for(let i = 0; i < 25; i++){
+    photos[i] = {};
+    photos[i].description = getRandomArrayElement(DESCRIPTIONS);
+    photos[i].comments = createComments();
+    photos[i].id = getRandomIdNumber();
+    photos[i].url = `photo/${getRandomUrlNumber()}.jpg`;
+    photos[i].likes = getRandomPositiveInteger(15,200);
+  }
+  return photos;
+};
+createPhotos();
